% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explode}
\alias{explode}
\title{Explode a colon-separated string into a data frame.
This function takes a character vector where each element contains
multiple values separated by colons (":") and splits them into a data frame.}
\usage{
explode(x)
}
\arguments{
\item{x}{A character vector with colon-separated values.}
}
\value{
A data frame with each column corresponding to a part of the colon-separated values.
}
\description{
Explode a colon-separated string into a data frame.
This function takes a character vector where each element contains
multiple values separated by colons (":") and splits them into a data frame.
}
\keyword{internal}
