% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.mic_solve}
\alias{tidy.mic_solve}
\title{tidy method for mic_solve objects}
\usage{
tidy.mic_solve(
  x,
  table = c("mic", "delta", "ratio", "DoD_delta", "DoD_ratio"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{mic_solve}.}

\item{table}{A character string specifying which table to return.
Options are \code{"mic"}, \code{"delta"}, \code{"ratio"}, \code{"DoD_delta"}, and \code{"DoD_ratio"}.}

\item{...}{Additional arguments (not used).}
}
\value{
A tibble containing the requested table from the \code{mic_solve} object.
}
\description{
tidy method for mic_solve objects
}
